add airbnb style guide
