+++ package.json
@@ -15,6 +15,11 @@
     "react-dom": "^15.5.4"
   },
   "devDependencies": {
+    "eslint": "^3.19.0",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-react": "^7.0.0",
     "react-scripts": "0.9.5"
   },
   "scripts": {
