+++ src/components/CommentBox.jsx
@@ -1,11 +1,11 @@
 // CommentBox
 
+import 'bulma/css/bulma.css';
+
 import React, { Component } from 'react';
 import axios from 'axios';
 import CommentList from './CommentList';
 import CommentForm from './CommentForm';
-
-import 'bulma/css/bulma.css';
 
 class CommentBox extends Component {
 
@@ -16,17 +16,21 @@ class CommentBox extends Component {
     this.loadCommentsFromServer = this.loadCommentsFromServer.bind(this);
     this.handleCommentSubmit = this.handleCommentSubmit.bind(this);
     this.handleCommentDelete = this.handleCommentDelete.bind(this);
+  }
 
+  componentDidMount() {
+    this.loadCommentsFromServer();
+    setInterval(this.loadCommentsFromServer, this.props.pollInterval);
   }
 
   loadCommentsFromServer() {
-    axios.get(this.props.url).then( res => {
+    axios.get(this.props.url).then(res => {
       this.setState({ data: res.data });
-    })
+    });
   }
 
   handleCommentSubmit(comment) {
-    let comments = this.state.data;
+    const comments = this.state.data;
     comment._id = Date.now();
     let newComments = comments.concat([comment]);
     this.setState({ data: newComments });
@@ -38,28 +42,17 @@ class CommentBox extends Component {
   }
 
   handleCommentDelete(id) {
-
-    let comments = this.state.data;
-
-    // comments.splice(0, 1);
-    let newComments = comments.filter( (t) => {
-      return t._id !== id 
+    const comments = this.state.data;
+    const newComments = comments.filter((t) => {
+      return t._id !== id
     });
 
     this.setState({ data: newComments });
 
     axios.delete(`${this.props.url}/${id}`)
-      .then(res => {
-        console.log('Comment deleted');
-      })
-      .catch( err => {
+      .catch(err => {
         console.error(err);
       })
-  }
-
-  componentDidMount() {
-    this.loadCommentsFromServer();
-    setInterval(this.loadCommentsFromServer, this.props.pollInterval);
   }
 
   render() {
@@ -67,14 +60,14 @@ class CommentBox extends Component {
       <section className="section">
         <div className="container">
 
-        <CommentList 
-          data={ this.state.data }
-          onCommentDelete={ this.handleCommentDelete }
-        />
-        <hr/>
-        <CommentForm 
-          onCommentSubmit={ this.handleCommentSubmit }
-        />
+          <CommentList
+            data={this.state.data}
+            onCommentDelete={this.handleCommentDelete}
+          />
+          <hr />
+          <CommentForm
+            onCommentSubmit={this.handleCommentSubmit}
+          />
         </div>
       </section>
     )
